+++ include/SemiGlobalMatching/CostCalculator/smNullPathCostAggregator.h
@@ -5,6 +5,6 @@ namespace SemiGlobalMatching {
     public:
 
     public:
-        void smCostAggregate(u8* imageData, u8* costMatrix, u32 imageWidth, u32 imageHeight,i32 minDisparity, u32 disparityRange, u32* refinedMatrix) override;
+        void smCostAggregate(u8* imageData, u32* costMatrix, u32 imageWidth, u32 imageHeight,i32 minDisparity, u32 disparityRange, u32* refinedMatrix) override;
     };
 }
